refactor: reorganize composables into feature-based subdirectories
## Summary

Migrates all composables from a flat `src/composables/` directory into
feature-based subdirectories with barrel `index.ts` files per directory.
Pure refactor — no logic changes.

## New Structure
```
src/composables/
  data/        — localStorage, job data, applications, visited sites, ATS
  ui/          — singleton dialog state and focus registry
  keyboard/    — shortcut registration and input guards
  lib/         — generic, reusable primitives (grouped options, select model)
  dashboard/   — view-specific derived data (category progress)
```

## Changes

### Composable moves
- `use-applications`, `use-ats-detection`, `use-data-management`,
  `use-job-data`, `use-job-sites`, `use-visited-sites` → `data/`
- `use-add-application-dialog`, `use-command-palette`,
  `use-shortcut-reference`, `use-site-focus` → `ui/`
- `use-input-guard`, `use-keyboard-shortcuts` → `keyboard/`
- `use-grouped-options`, `use-select-model` → `lib/`
- `use-category-progress` → `dashboard/`

### Barrel files
- Added `index.ts` per subdirectory using `./relative` imports
- Consumers import from subdirectory: `@/composables/data`

### Import convention enforced
- Intra-directory imports use `./relative` paths to avoid circular
  barrel initialization (fixes `useJobSites is not a function` in tests)
- Cross-directory imports use `@/` alias
- All `vi.mock()` paths in test files updated to match new locations

### Updated consumers
- `src/App.vue`
- `src/views/Home.vue`, `Applications.vue`
- `src/components/app/applications/add-application-dialog/AddApplicationDialog.vue`
- `src/components/app/lib/BaseSelect.vue`
- `src/components/app/shell/command-palette/CommandPalette.vue`,
  `CommandPaletteSites.vue`
- `src/components/app/shell/shortcut-reference-dialog/ShortcutReferenceDialog.vue`
- `src/components/app/sites/job-site-card/JobSiteCard.vue`
- All affected test files

### Documentation
- Updated `src/composables/README.md` to reflect new structure,
  directory purposes, and import convention

## Verification

- `npm run lint` — 0 warnings
- `npm run format:check` — clean
- `npm run test` — 395/395 tests passing across 30 test files

Closes #22
