+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "job-hunt-daily",
-  "version": "1.3.0",
+  "version": "1.3.1",
   "scripts": {
     "dev": "vite",
     "build": "vue-tsc -b && vite build",
