+++ src/composables/dashboard/index.ts
@@ -1,0 +1,3 @@
+// /src/composables/dashboard/index.ts
+
+export { useCategoryProgress } from "./use-category-progress";
