+++ src/composables/data/use-applications.ts
@@ -1,4 +1,4 @@
-// /src/composables/use-applications.ts
+// /src/composables/data/use-applications.ts
 
 import { useLocalStorage } from "@vueuse/core";
 import { computed } from "vue";
