+++ src/composables/data/use-ats-detection.test.ts
@@ -1,10 +1,11 @@
-// /src/composables/use-ats-detection.test.ts
+// /src/composables/data/use-ats-detection.test.ts
 
 import { describe, expect, it } from "vitest";
 
-import { useATSDetection } from "@/composables/use-ats-detection";
 import { mockJobHuntData } from "@/test-utils/mocks";
 import type { JobHuntData, JobSite } from "@/types";
+
+import { useATSDetection } from "./use-ats-detection";
 
 describe("useATSDetection", () => {
   describe("getATS", () => {
