+++ src/composables/data/use-visited-sites.test.ts
@@ -1,12 +1,13 @@
-// /src/composables/use-visited-sites.test.ts
+// /src/composables/data/use-visited-sites.test.ts
 
 import { describe, it, expect, beforeEach, vi } from "vitest";
 import { nextTick } from "vue";
 
 import { TEST_VISITED_SITES_STORAGE_KEY } from "@/composables/keys";
-import { useVisitedSites } from "@/composables/use-visited-sites";
 import { getToday, todayIso } from "@/lib/time";
 import { withFrozenTime } from "@/test-utils/with-frozen-time";
+
+import { useVisitedSites } from "./use-visited-sites";
 
 describe("useVisitedSites", () => {
   beforeEach(() => {
