+++ src/composables/lib/index.ts
@@ -1,0 +1,6 @@
+// /src/composables/lib/index.ts
+
+export { useGroupedOptions } from "./use-grouped-options";
+export type { GroupableOption } from "./use-grouped-options";
+
+export { useSelectModel } from "./use-select-model";
